+++ Locale/zhCN.lua
@@ -217,6 +217,7 @@ local SimplifiedChinese = {
   ["Pin to top"] = "置顶",
   ["Unpin"] = "取消置顶",
   ["Remove"] = "移除",
+  ["Mark last messages as unread"] = "将最近消息标记为未读",
   ["Copy text"] = "复制文本",
   ["Search chats"] = "搜索对话",
   ["Select a conversation or"] = "选择一个对话或",